requirements.txt
bcrypt==3.2.0
cffi==1.14.4
pycparser==2.20
pycrypto==2.6.1
six==1.15.0
